requirements.txt
flask
openai
transformers
torch
sentencepiece
python-dotenv